+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Render custom emojis in poll options.
+- Navigate to parent post when clicking on "replying to".
 
 ### Changed